# controller/swagger/swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/motorista": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Motorista"
        ],
        "operationId": "listMotorista",
        "parameters": [
          {
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListMotorista"
            }
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Motorista"
        ],
        "operationId": "postMotorista",
        "parameters": [
          {
            "in": "body",
            "schema": {
              "$ref": "#/definitions/InsertMotorista"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ReturnID"
            }
          }
        }
      }
    },
    "/motorista/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Motorista"
        ],
        "operationId": "getMotorista",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Motorista"
            }
          }
        }
      },
      "put": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Motorista"
        ],
        "operationId": "putMotorista",
        "parameters": [
          {
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateMotorista"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ReturnID"
            }
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Motorista"
        ],
        "operationId": "deleteMotorista",
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/veiculo": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Veiculo"
        ],
        "operationId": "listVeiculo",
        "parameters": [
          {
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListVeiculo"
            }
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Veiculo"
        ],
        "operationId": "postVeiculo",
        "parameters": [
          {
            "in": "body",
            "schema": {
              "$ref": "#/definitions/InsertVeiculo"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ReturnID"
            }
          }
        }
      }
    },
    "/veiculo/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Veiculo"
        ],
        "operationId": "getVeiculo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Veiculo"
            }
          }
        }
      },
      "put": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Veiculo"
        ],
        "operationId": "putVeiculo",
        "parameters": [
          {
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateVeiculo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ReturnID"
            }
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Veiculo"
        ],
        "operationId": "deleteVeiculo",
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
...
    "InsertMotorista": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "x-go-name": "Nome"
        }
      },
      "x-go-package": "transportadora/models"
    },
    "InsertVeiculo": {
      "type": "object",
      "properties": {
        "placa": {
          "type": "string",
          "x-go-name": "Placa"
        }
      },
      "x-go-package": "transportadora/models"
    },
    "ListMotorista": {
      "type": "object",
      "properties": {
        "Data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Motorista"
          }
        },
        "Pagination": {
          "$ref": "#/definitions/PaginationResponse"
        }
      },
      "x-go-package": "transportadora/models"
    },
...
    "ListVeiculo": {
      "type": "object",
      "properties": {
        "Data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Veiculo"
          }
        },
        "Pagination": {
          "$ref": "#/definitions/PaginationResponse"
        }
      },
      "x-go-package": "transportadora/models"
    },
...
    "Motorista": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "nome": {
          "type": "string",
          "x-go-name": "Nome"
        },
        "veiculo": {
          "$ref": "#/definitions/Veiculo"
        }
      },
      "x-go-package": "transportadora/models"
    },
...
    "PaginationResponse": {
      "type": "object",
      "properties": {
        "current_page": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CurrentPage"
        },
        "record_per_page": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RecordPerPage"
        },
        "total_page": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TotalPage"
        },
        "total_record": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TotalRecord"
        }
      },
      "x-go-package": "transportadora/models"
    },
...
    "ReturnID": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        }
      },
      "x-go-package": "transportadora/models"
    },
...
    "UpdateMotorista": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "x-go-name": "Nome"
        }
      },
      "x-go-package": "transportadora/models"
    },
    "UpdateVeiculo": {
      "type": "object",
      "properties": {
        "placa": {
          "type": "string",
          "x-go-name": "Placa"
        }
      },
      "x-go-package": "transportadora/models"
    },
...
    "Veiculo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "placa": {
          "type": "string",
          "x-go-name": "Placa"
        }
      },
      "x-go-package": "transportadora/models"
    },
...
  }
}