# infra/pgstore/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package pgstore

import (
	"github.com/google/uuid"
)

type Motorista struct {
	ID        uuid.UUID `db:"id"`
	VeiculoID uuid.UUID `db:"veiculo_id"`
	Nome      string    `db:"nome"`
}

type Veiculo struct {
	ID    uuid.UUID `db:"id"`
	Placa string    `db:"placa"`
}